+++ CHANGELOG.md
@@ -9,5 +9,7 @@
 - Add update() method to Account
 - Add delete() method to Account
 - Add move() method to Account
+- Add reorder() method to Account
 - Add proper testing for ToshlException
 - Improve string representation of ToshlException
+- refactoring: each entity is in a separate module